test bootstrap and add ability,samepa rankng,usage

+++ src/main/java/tripleusage/com/usage/domain/PokemonTotal.java
@@ -2,5 +2,5 @@ package tripleusage.com.usage.domain;
 
 public interface PokemonTotal {
     String getName();
-    int getPokemontotal();
+    float getPokemontotal();
 }
